add shape layer
