+++ pyspark/bigdl/nn/onnx/layer.py
@@ -1,0 +1,33 @@
+#
+# Copyright 2016 The BigDL Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import sys
+from bigdl.nn.layer import Layer
+
+if sys.version >= '3':
+    long = int
+    unicode = str
+
+
+class Shape(Layer):
+    """
+    A layer which takes a tensor as input and outputs an 1D tensor containing the shape of the input.
+
+    >>> shape = Shape()
+    creating: createShape
+    """
+    def __init__(self, bigdl_type="float"):
+        super(Shape, self).__init__(None, bigdl_type)
